+++ RELEASE_NOTES.md
@@ -141,3 +141,4 @@
 #### 2.0.0-alpha7 - Unreleased
 * Improved name generation algorithm to cope better with acronymns.
 * Fixed wrong singularization of words ending with 'uses'.
+* Fixed handling of repeated one letter names.