+++ RELEASE_NOTES.md
@@ -146,3 +146,4 @@
 * Support for more api patterns in ApiaryProvider.
 * Tolerate invalid json and missing data in apiary specifications.
 * Added 'SpecialNames' parameter to ApiaryProvider to allow taking concrete examples in the apiary specification and turning them to api parameters.
+* Improved naming of generated types.